+++ src/main/java/org/folio/circulation/resources/RequestNoticeSender.java
@@ -1,5 +1,6 @@
 package org.folio.circulation.resources;
 
+import static java.util.concurrent.CompletableFuture.runAsync;
 import static org.folio.circulation.domain.notice.TemplateContextUtil.createRequestNoticeContext;
 
 import java.util.Collections;
@@ -11,6 +12,7 @@ import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.Request;
 import org.folio.circulation.domain.RequestAndRelatedRecords;
 import org.folio.circulation.domain.representations.logs.NoticeLogContext;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.infrastructure.storage.requests.RequestRepository;
 import org.folio.circulation.domain.RequestStatus;
 import org.folio.circulation.domain.RequestType;
@@ -38,19 +40,22 @@ public class RequestNoticeSender {
 
   public static RequestNoticeSender using(Clients clients) {
     return new RequestNoticeSender(PatronNoticeService.using(clients),
-      RequestRepository.using(clients),
+      RequestRepository.using(clients), new LoanRepository(clients),
       new EventPublisher(clients.pubSubPublishingService()));
   }
 
   private final PatronNoticeService patronNoticeService;
   private final RequestRepository requestRepository;
+  private final LoanRepository loanRepository;
   private final EventPublisher eventPublisher;
 
   public RequestNoticeSender(PatronNoticeService patronNoticeService,
                              RequestRepository requestRepository,
+                             LoanRepository loanRepository,
                              EventPublisher eventPublisher) {
     this.patronNoticeService = patronNoticeService;
     this.requestRepository = requestRepository;
+    this.loanRepository = loanRepository;
     this.eventPublisher = eventPublisher;
   }
 
@@ -74,8 +79,15 @@ public class RequestNoticeSender {
     Loan loan = request.getLoan();
     if (request.getRequestType() == RequestType.RECALL && loan != null) {
       sendNoticeOnItemRecalledEvent(loan);
+      sendLogEvent(requester, loan);
     }
     return Result.succeeded(relatedRecords);
+  }
+
+  private void sendLogEvent(User user, Loan loan) {
+    runAsync(() -> loanRepository.getById(loan.getId())
+      .thenAccept(existingLoan -> existingLoan.map(l -> loan.setPreviousDueDate(l.getDueDate())))
+      .thenApply(vVoid -> eventPublisher.publishRecallRequestedEvent(loan.copy().withUser(user))));
   }
 
   public Result<RequestAndRelatedRecords> sendNoticeOnRequestMoved(
@@ -86,6 +98,7 @@ public class RequestNoticeSender {
     Loan loan = request.getLoan();
     if (request.getRequestType() == RequestType.RECALL && loan != null) {
       sendNoticeOnItemRecalledEvent(loan);
+      sendLogEvent(request.getRequester(), loan);
     }
     return Result.succeeded(relatedRecords);
   }
@@ -124,8 +137,7 @@ public class RequestNoticeSender {
         .withEventType(NoticeEventType.ITEM_RECALLED)
         .withNoticeContext(TemplateContextUtil.createLoanNoticeContext(loan))
         .build();
-      patronNoticeService.acceptNoticeEvent(itemRecalledEvent, NoticeLogContext.from(loan))
-        .thenCompose(r -> r.after(v -> eventPublisher.publishRecallRequestedEvent(loan)));
+      patronNoticeService.acceptNoticeEvent(itemRecalledEvent, NoticeLogContext.from(loan));
     }
     return Result.succeeded(null);
   }