+++ src/main/java/org/folio/circulation/services/EventPublisher.java
@@ -1,6 +1,7 @@
 package org.folio.circulation.services;
 
 import static java.util.concurrent.CompletableFuture.completedFuture;
+import static java.util.concurrent.CompletableFuture.runAsync;
 import static org.folio.circulation.domain.EventType.ITEM_AGED_TO_LOST;
 import static org.folio.circulation.domain.EventType.ITEM_CHECKED_IN;
 import static org.folio.circulation.domain.EventType.ITEM_CHECKED_OUT;
@@ -9,11 +10,14 @@ import static org.folio.circulation.domain.EventType.ITEM_DECLARED_LOST;
 import static org.folio.circulation.domain.EventType.LOAN_DUE_DATE_CHANGED;
 import static org.folio.circulation.domain.EventType.LOG_RECORD;
 import static org.folio.circulation.domain.LoanAction.CHECKED_IN;
+import static org.folio.circulation.domain.LoanAction.DUE_DATE_CHANGED;
+import static org.folio.circulation.domain.LoanAction.RECALLREQUESTED;
 import static org.folio.circulation.domain.representations.logs.LogEventPayloadField.LOG_EVENT_TYPE;
-import static org.folio.circulation.domain.representations.logs.CirculationCheckInCheckOutLogEventMapper.mapToCheckInLogEventJson;
-import static org.folio.circulation.domain.representations.logs.CirculationCheckInCheckOutLogEventMapper.mapToCheckOutLogEventJson;
+import static org.folio.circulation.domain.representations.logs.CirculationCheckInCheckOutLogEventMapper.mapToCheckInLogEventContent;
+import static org.folio.circulation.domain.representations.logs.CirculationCheckInCheckOutLogEventMapper.mapToCheckOutLogEventContent;
 import static org.folio.circulation.domain.representations.logs.LogEventPayloadField.PAYLOAD;
 import static org.folio.circulation.domain.representations.logs.LogEventType.LOAN;
+import static org.folio.circulation.domain.representations.logs.LogEventType.NOTICE;
 import static org.folio.circulation.support.AsyncCoordinationUtil.allOf;
 import static org.folio.circulation.domain.representations.logs.RequestUpdateLogEventMapper.mapToRequestLogEventJson;
 import static org.folio.circulation.support.json.JsonPropertyWriter.write;
@@ -28,15 +32,21 @@ import org.folio.circulation.domain.EventType;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.LoanAndRelatedRecords;
 import org.folio.circulation.domain.RequestAndRelatedRecords;
+import org.folio.circulation.domain.User;
 import org.folio.circulation.domain.anonymization.LoanAnonymizationRecords;
 import org.folio.circulation.domain.representations.logs.LoanLogContext;
 import org.folio.circulation.domain.representations.logs.LogContextActionResolver;
 import org.folio.circulation.domain.representations.logs.LogEventType;
 import org.folio.circulation.domain.Request;
+import org.folio.circulation.domain.representations.logs.NoticeLogContext;
 import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
+import org.folio.circulation.infrastructure.storage.users.UserRepository;
 import org.folio.circulation.resources.context.RenewalContext;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.results.Result;
+import org.joda.time.DateTime;
+import org.joda.time.format.DateTimeFormat;
+import org.joda.time.format.DateTimeFormatter;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -50,6 +60,7 @@ public class EventPublisher {
   public static final String DUE_DATE_CHANGED_BY_RECALL_FIELD = "dueDateChangedByRecall";
   public static final String FAILED_TO_PUBLISH_LOG_TEMPLATE =
     "Failed to publish {} event: loan is null";
+  public static final DateTimeFormatter FMT = DateTimeFormat.forPattern("yyyy-MM-dd HH:mm:ss");
 
   private final PubSubPublishingService pubSubPublishingService;
 
@@ -62,7 +73,7 @@ public class EventPublisher {
   }
 
   public CompletableFuture<Result<LoanAndRelatedRecords>> publishItemCheckedOutEvent(
-    LoanAndRelatedRecords loanAndRelatedRecords) {
+    LoanAndRelatedRecords loanAndRelatedRecords, UserRepository userRepository) {
 
     if (loanAndRelatedRecords.getLoan() != null) {
       Loan loan = loanAndRelatedRecords.getLoan();
@@ -72,8 +83,9 @@ public class EventPublisher {
       write(payloadJsonObject, LOAN_ID_FIELD, loan.getId());
       write(payloadJsonObject, DUE_DATE_FIELD, loan.getDueDate());
 
-      JsonObject logEventPayload = mapToCheckOutLogEventJson(loanAndRelatedRecords);
-      CompletableFuture.runAsync(() -> pubSubPublishingService.publishEvent(LOG_RECORD.name(), logEventPayload.encode()));
+      runAsync(() -> userRepository.getUser(loanAndRelatedRecords.getLoggedInUserId())
+        .thenApplyAsync(r -> r.after(loggedInUser -> CompletableFuture.completedFuture(
+          Result.succeeded(pubSubPublishingService.publishEvent(LOG_RECORD.name(), mapToCheckOutLogEventContent(loanAndRelatedRecords, loggedInUser)))))));
 
       return pubSubPublishingService.publishEvent(ITEM_CHECKED_OUT.name(), payloadJsonObject.encode())
         .thenApply(r -> succeeded(loanAndRelatedRecords));
@@ -86,10 +98,11 @@ public class EventPublisher {
   }
 
   public CompletableFuture<Result<CheckInContext>> publishItemCheckedInEvents(
-    CheckInContext checkInContext) {
+    CheckInContext checkInContext, UserRepository userRepository) {
 
-    JsonObject logEventPayload = mapToCheckInLogEventJson(checkInContext);
-    CompletableFuture.runAsync(() -> pubSubPublishingService.publishEvent(LOG_RECORD.name(), logEventPayload.encode()));
+    runAsync(() -> userRepository.getUser(checkInContext.getLoggedInUserId())
+      .thenApplyAsync(r -> r.after(loggedInUser -> CompletableFuture.completedFuture(
+        Result.succeeded(pubSubPublishingService.publishEvent(LOG_RECORD.name(), mapToCheckInLogEventContent(checkInContext, loggedInUser)))))));
 
     if (checkInContext.getLoan() != null) {
       Loan loan = checkInContext.getLoan();
@@ -129,7 +142,7 @@ public class EventPublisher {
 
     LoanLogContext loanLogContext = LoanLogContext.from(loan)
       .withDescription(String.format("Additional information: %s", loan.getActionComment()));
-    CompletableFuture.runAsync(() -> publishLogRecord(loanLogContext.asJson(), LOAN));
+    runAsync(() -> publishLogRecord(loanLogContext.asJson(), LOAN));
 
     JsonObject payloadJson = new JsonObject();
     write(payloadJson, USER_ID_FIELD, loan.getUserId());
@@ -139,7 +152,11 @@ public class EventPublisher {
       .thenApply(r -> succeeded(loan));
   }
 
-  private CompletableFuture<Result<Loan>> publishDueDateChangedEvent(Loan loan) {
+  private CompletableFuture<Result<Loan>> publishDueDateChangedEvent(Loan loan, RequestAndRelatedRecords records) {
+    return publishDueDateChangedEvent(loan, records.getRequest().getRequester());
+  }
+
+  private CompletableFuture<Result<Loan>> publishDueDateChangedEvent(Loan loan, User user) {
     if (loan != null) {
       JsonObject payloadJsonObject = new JsonObject();
       write(payloadJsonObject, USER_ID_FIELD, loan.getUserId());
@@ -147,11 +164,7 @@ public class EventPublisher {
       write(payloadJsonObject, DUE_DATE_FIELD, loan.getDueDate());
       write(payloadJsonObject, DUE_DATE_CHANGED_BY_RECALL_FIELD, loan.wasDueDateChangedByRecall());
 
-      LoanLogContext loanLogContext = LoanLogContext.from(loan)
-        .withAction(LogContextActionResolver.resolveAction(loan.getAction()))
-        .withDescription(String.format("New due date: %s (from %s)",
-          loan.getDueDate(), loan.getOriginalDueDate()));
-      CompletableFuture.runAsync(() -> publishLogRecord(loanLogContext.asJson(), LOAN));
+      runAsync(() -> publishDueDateLogEvent(loan.copy().withUser(user)));
 
       return pubSubPublishingService.publishEvent(LOAN_DUE_DATE_CHANGED.name(),
         payloadJsonObject.encode())
@@ -169,7 +182,7 @@ public class EventPublisher {
 
     if (loanAndRelatedRecords.getLoan() != null) {
       Loan loan = loanAndRelatedRecords.getLoan();
-      publishDueDateChangedEvent(loan);
+      publishDueDateChangedEvent(loan, loan.getUser());
     }
 
     return completedFuture(succeeded(loanAndRelatedRecords));
@@ -178,7 +191,9 @@ public class EventPublisher {
   public CompletableFuture<Result<RenewalContext>> publishDueDateChangedEvent(
     RenewalContext renewalContext) {
 
-    publishDueDateChangedEvent(renewalContext.getLoan());
+    var loan = renewalContext.getLoan();
+
+    publishDueDateChangedEvent(loan, loan.getUser());
 
     return completedFuture(succeeded(renewalContext));
   }
@@ -188,7 +203,7 @@ public class EventPublisher {
 
     LoanRepository loanRepository = new LoanRepository(clients);
     loanRepository.findOpenLoanForRequest(requestAndRelatedRecords.getRequest())
-      .thenCompose(r -> r.after(this::publishDueDateChangedEvent));
+      .thenCompose(r -> r.after(loan -> publishDueDateChangedEvent(loan, requestAndRelatedRecords)));
 
     return completedFuture(succeeded(requestAndRelatedRecords));
   }
@@ -225,8 +240,17 @@ public class EventPublisher {
 
   public CompletableFuture<Result<Void>> publishRecallRequestedEvent(Loan loan) {
     return publishLogRecord(LoanLogContext.from(loan)
-      .withAction("Recall requested")
-      .withDescription(String.format("New due date: %s (from %s)", loan.getDueDate(), loan.getOriginalDueDate())).asJson(), LOAN);
+      .withAction(LogContextActionResolver.resolveAction(RECALLREQUESTED.getValue()))
+      .withDescription(String.format("New due date: %s (from %s)", formatDateTime(loan.getDueDate()), formatDateTime(loan.getPreviousDueDate()))).asJson(), LOAN);
+  }
+  public CompletableFuture<Result<Void>> publishDueDateLogEvent(Loan loan) {
+    return publishLogRecord(LoanLogContext.from(loan)
+      .withAction(LogContextActionResolver.resolveAction(DUE_DATE_CHANGED.getValue()))
+      .withDescription(String.format("New due date: %s (from %s)", formatDateTime(loan.getDueDate()), formatDateTime(loan.getPreviousDueDate()))).asJson(), LOAN);
+  }
+
+  public CompletableFuture<Result<Void>> publishNoticeEvent(NoticeLogContext noticeLogContext) {
+    return publishLogRecord(noticeLogContext.withDate(DateTime.now()).asJson(), NOTICE);
   }
 
   public CompletableFuture<Result<Void>> publishLogRecord(JsonObject context, LogEventType payloadType) {
@@ -238,7 +262,11 @@ public class EventPublisher {
   }
 
   public RequestAndRelatedRecords publishLogRecordAsync(RequestAndRelatedRecords requestAndRelatedRecords, Request originalRequest, LogEventType logEventType) {
-    CompletableFuture.runAsync(() -> publishLogRecord(mapToRequestLogEventJson(originalRequest, requestAndRelatedRecords.getRequest()), logEventType));
+    runAsync(() -> publishLogRecord(mapToRequestLogEventJson(originalRequest, requestAndRelatedRecords.getRequest()), logEventType));
     return requestAndRelatedRecords;
+  }
+
+  private String formatDateTime(DateTime dateTime) {
+    return dateTime.toString(FMT);
   }
 }