+++ src/test/java/api/loans/agetolost/ScheduledAgeToLostFeeChargingApiTest.java
@@ -355,6 +355,30 @@ public class ScheduledAgeToLostFeeChargingApiTest extends SpringApiTest {
   }
 
   @Test
+  public void shouldAgeToLostAndChargeLostItemProcessingFeeWhenActualFeeSet() {
+    final double agedToLostLostProcessingFee = 10.00;
+    IndividualResource lostItemPolicy = lostItemFeePoliciesFixture.create(
+      lostItemFeePoliciesFixture.ageToLostAfterOneMinutePolicy()
+        .doNotChargeOverdueFineWhenReturned()
+        .withNoFeeRefundInterval()
+        .withActualCost(10.00)
+        .billPatronImmediatelyWhenAgedToLost()
+        .chargeProcessingFeeWhenAgedToLost(agedToLostLostProcessingFee));
+
+    useLostItemPolicy(lostItemPolicy.getId());
+    
+    final ItemResource item = itemsFixture.basedUponNod();
+    final CheckOutResource checkOut = checkOutFixture.checkOutByBarcode(item);
+
+    ageToLostFixture.ageToLostAndChargeFees();
+
+    final IndividualResource loanFromStorage = loansStorageClient.get(checkOut.getId());
+
+    assertThat(itemsFixture.getById(item.getId()).getJson(), isAgedToLost());
+    assertThat(loanFromStorage, hasLostItemProcessingFee(isOpen(agedToLostLostProcessingFee)));
+  }
+
+  @Test
   public void declaredLostItemShouldNotBeAgedToLost() {
     final double declaredLostProcessingFee = 10.00;
     useLostItemPolicy(lostItemFeePoliciesFixture.create(